Create First match algorithm OSCR and CH
This exercise aims to match the organisations on OSCR's publicly available dataset which identify as companies to the Companies House register via their API. This first stage algorithm matches on name and postcode. A namescore of 100 is accepted as a match, if namescore <100, postcode score is given a weight of 0.5. Overall, a combined score threshold of 78 is accepted as a match, all else is declared non-match.