Create Second stage match OSCR and CH
To refine the matching process, post manual accuracy checking, the original algorithm (documented in 'First match OSCR and CH')  was adjusted to return multiple (in this case the top 5) matches from Companies House. All other elements (threshold for matching and postcode weighting remain the same. By doing so, it is possible to see where a match has been accurately made further down the list of returns as the original algorithm only returns the single ‘best’ match. In addition, information on names companies were ‘also known as’ (where available) was included within the OSCR dataset to be matched and the alternative name held by Companies House was included within the matching process.