Rename companydirectorloop to Company loop on CH API to identify directors
This program builds on the previous (documented in 'Individual company directors identifier') to loop through a list of companies and read the relevant information on all directors for each to an output file. It currently works but only identifies a single director from each company and requires further refinement to identify all associated directors of each company.