Rename First match algorithm OSCR and CH to Companies Algorithm 1 - probabilistic matching between OSCR and CHAPI data
The first probabilistic matching algorithm designed to take the name and postcode of the charity from the Scottish Charity Register and compare them with company names and postcodes within CHAPI. Identical names between the two datasets are automatically accepted as a match (namescore = 100). Where there are differences identified in names, the score assigned to the name comparison was combined with a score derived from a comparison of the postcodes recorded on each dataset, with a weight of 0.5 assigned to the latter.