Rename Second stage match OSCR and CH to Companies Algorithm 2 - improving the probabilistic match between OSCR and CHAPI data
A refinement of Companies Algorithm 1, designed to expand the number of potential match organisations returned by CHAPI to increase the likelihood of obtaining a match. Company algorithm 1 was adjusted to return multiple (in this case the top 5 scoring) matches from Companies House data. All other elements (threshold for matching and postcode weighting) remain the same.